requirements.txt
# requirements.txt
streamlit
requests
pandas
python-dotenv